+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.3.0
+- Add Apotheosis World Tier support to rendering
+- Fix damage modifiers
+- Change damage-based modifiers to be addition-based (using new config structure - regenerate your common config!)
+- Various cleanup
+
 ## 0.2.0
 - Fix attribute modifiers not applying correctly